refactor: 更新日志手动解析出可点击链接 (#120)

+++ src/components/Update/index.tsx
@@ -63,10 +63,12 @@ const Update = () => {
 		try {
 			const { shouldUpdate, manifest } = await tauriCheckUpdate();
 
-			if (shouldUpdate) {
+			if (shouldUpdate && manifest) {
 				showWindow();
 
 				messageApi.destroy(MESSAGE_KEY);
+
+				manifest.body = replaceManifestBody(manifest.body);
 
 				Object.assign(state, { manifest, open: true });
 			} else if (showMessage) {
@@ -85,6 +87,19 @@ const Update = () => {
 				content: "检查更新时出错，请检查网络并重试。",
 			});
 		}
+	};
+
+	const replaceManifestBody = (body: string) => {
+		return (
+			body
+				// 替换贡献者名称
+				.replace(
+					/(-.*?by.*?)@([^ ]+)/g,
+					"$1<a href='https://github.com/$2'><mark>@$2</mark></a>",
+				)
+				// 替换 pr 链接
+				.replace(new RegExp(`(${GITHUB_ISSUES_LINK}/)(\\d+)`), "[#$2]($1$2)")
+		);
 	};
 
 	const handleOk = async () => {
@@ -140,6 +155,7 @@ const Update = () => {
 				title="发现新版本🥳"
 				okText="立即更新"
 				cancelText="以后再说"
+				className={styles.modal}
 				confirmLoading={state.loading}
 				onOk={handleOk}
 				onCancel={handleCancel}
@@ -149,7 +165,9 @@ const Update = () => {
 						更新版本：
 						<span>
 							v{appInfo?.version} 👉{" "}
-							<span className="text-primary">v{state.manifest?.version}</span>
+							<a href={`${GITHUB_LINK}/releases/latest`}>
+								v{state.manifest?.version}
+							</a>
 						</span>
 					</Flex>
